+++ apps/dataset-browser/package.json
@@ -20,6 +20,7 @@
     "test:e2e:open": "cypress open"
   },
   "dependencies": {
+    "@colonial-collections/iris": "*",
     "@floating-ui/react": "0.23.0",
     "@hapi/hoek": "11.0.2",
     "@headlessui/react": "1.7.14",
