+++ package.json
@@ -14,6 +14,7 @@
     "fix": "gts fix",
     "lint": "gts lint",
     "test": "turbo run test",
+    "test:watch": "turbo run test:watch",
     "test:ci": "turbo run test:ci",
     "test:integration": "turbo run test:integration",
     "test:e2e": "turbo run test:e2e"
