+++ packages/iris/package.json
@@ -1,0 +1,42 @@
+{
+  "name": "@colonial-collections/iris",
+  "version": "1.0.0",
+  "private": true,
+  "engines": {
+    "node": "18.x"
+  },
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "license": "MIT",
+  "scripts": {
+    "clean": "gts clean",
+    "compile": "tsc",
+    "fix": "gts fix",
+    "lint": "gts check",
+    "test": "jest",
+    "test:ci": "npm test -- --ci",
+    "test:watch": "npm test -- --watchAll"
+  },
+  "dependencies": {
+    "validate-iri": "1.0.1",
+    "zod": "3.21.4"
+  },
+  "devDependencies": {
+    "@jest/globals": "29.5.0",
+    "@types/jest": "29.5.0",
+    "@types/node": "18.15.11",
+    "eslint-config-custom": "*",
+    "gts": "3.1.1",
+    "jest": "29.5.0",
+    "ts-jest": "29.1.0",
+    "tsconfig": "*",
+    "typescript": "5.0.4"
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "testTimeout": 60000,
+    "testMatch": ["<rootDir>/src/**/*.test.ts"],
+    "collectCoverage": true,
+    "collectCoverageFrom": ["<rootDir>/src/**/*.ts"]
+  }
+}
