+++ packages/iris/src/get-iris.ts
@@ -1,18 +1,4 @@
-import {validateIri, IriValidationStrategy} from 'validate-iri';
-import {z} from 'zod';
-
-export function isIri(value: string) {
-  const result = z
-    .string()
-    // Zod's native url() validator is not RFC 3987-compliant
-    .refine((arg: string) => {
-      const error = validateIri(arg, IriValidationStrategy.Pragmatic);
-      return error === undefined;
-    })
-    .safeParse(value);
-
-  return result.success;
-}
+import {isIri} from '.';
 
 // Gets all RFC 3987-compliant IRIs from an object, regardless of the object's nesting
 export function getIrisFromObject<T>(rootObject: T) {
