+++ packages/iris/src/index.ts
@@ -1,0 +1,2 @@
+export * from './get-iris';
+export * from './is-iri';
